Allow useRestQuery to work with normal promises
