+++ ui/apps/platform/src/Containers/Dashboard/hooks/useAlertGroups.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { AlertQueryGroupBy, fetchSummaryAlertCounts } from 'services/AlertsService';
-import useRestQuery from './useRestQuery';
+import useRestQuery from 'hooks/useRestQuery';
 
 export default function useAlertGroups(query: string, groupBy?: AlertQueryGroupBy) {
     const restQuery = useCallback(
