+++ ui/apps/platform/src/Containers/Dashboard/hooks/useDeploymentsAtRisk.ts
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { fetchDeploymentsWithProcessInfo } from 'services/DeploymentsService';
 import { SearchFilter } from 'types/search';
-import useRestQuery from './useRestQuery';
+import useRestQuery from 'hooks/useRestQuery';
 
 export default function useDeploymentsAtRisk(searchFilter: SearchFilter, numberOfResults = 6) {
     const restQuery = useCallback(() => {
